выделила в отдельный модуль функцию createPhotos
